+++ src/main/java/com/Core/Serwiss.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public class Serwiss {
     static final String FALLBACK_NAME = "world";
-    static final Lang FALLBACK_LANG = new Lang(1L, "hello", "en");
+    static final Lang FALLBACK_LANG = new Lang(1, "hello", "en");
     private final Logger logger = LoggerFactory.getLogger(Serwiss.class);
     private LangRepository repository;
 
@@ -21,9 +21,9 @@ public class Serwiss {
 
 
     String prepGreeting(String name, String lang) {
-        Long langId;
+        Integer langId;
         try{
-            langId = Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId());
+            langId = Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId());
         }catch (NumberFormatException e){
 
         logger.warn ("Non numeric language id used: " + lang);
